Fix latent crash paths and add fail-fast non-finite guards
- solver.py: repair two broken exception-recovery blocks around
  get_index_to_remove (missing 'as e', invalid arguments) so the
  restart-on-singularity path works instead of raising NameError/TypeError
- params.py: check_float converts with float() instead of int(), avoiding
  silent truncation of non-float numerics (e.g. np.float32)
- manager.py: guard the reduction-ratio denominator against 0/0 (clamp
  to the worst case)
- unconditional np.isfinite checks on objective and surrogate values;
  non-finite values exit with INVALID_EVAL_ERROR instead of silently
  corrupting the interpolation system

+++ upoqa/utils/params.py
@@ -486,7 +486,7 @@ def check_float(
     else:
         try:
             if val not in [np.inf, -np.inf]:
-                converted_val = int(val)  # Try converting val to an float
+                converted_val = float(val)  # Try converting val to a float
             else:
                 converted_val = val
             return (lower is None or converted_val >= lower) and (
